Limit maximum number of mutations with MaxBatch (#1068)
* Add BatchSize and Count to HighWatermark
* Add BatchSize to ReadLog
* Implement HighWatermarks inside sequencer

+++ core/sequencer/sequencer.go
@@ -47,13 +47,13 @@ func New(
 	mapAdmin tpb.TrillianAdminClient,
 	directories directory.Storage,
 	mutations mutator.MutationStorage,
-	batchSize int,
+	batchSize int32,
 ) *Sequencer {
 	return &Sequencer{
 		sequencerClient: sequencerClient,
 		directories:     directories,
 		mapAdmin:        mapAdmin,
-		batchSize:       int32(batchSize),
+		batchSize:       batchSize,
 	}
 }
